Move optimizers/two_qubit_state_preparation.py to transformers/analytical_decompositions/two_qubit_state_preparation.py
