+++ cirq-core/cirq/__init__.py
@@ -345,8 +345,6 @@ from cirq.optimizers import (
     MergeInteractions,
     MergeInteractionsToSqrtIswap,
     MergeSingleQubitGates,
-    prepare_two_qubit_state_using_cz,
-    prepare_two_qubit_state_using_sqrt_iswap,
     single_qubit_matrix_to_gates,
     single_qubit_matrix_to_pauli_rotations,
     single_qubit_matrix_to_phased_x_z,
@@ -366,6 +364,8 @@ from cirq.transformers import (
     map_operations_and_unroll,
     merge_moments,
     merge_operations,
+    prepare_two_qubit_state_using_cz,
+    prepare_two_qubit_state_using_sqrt_iswap,
     unroll_circuit_op,
     unroll_circuit_op_greedy_earliest,
     unroll_circuit_op_greedy_frontier,
