+++ cirq-core/cirq/transformers/analytical_decompositions/__init__.py
@@ -12,20 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Circuit transformation utilities."""
+"""Utilities for analytical decomposition of cirq gates."""
 
-from cirq.transformers.analytical_decompositions import (
+from cirq.transformers.analytical_decompositions.two_qubit_state_preparation import (
     prepare_two_qubit_state_using_cz,
     prepare_two_qubit_state_using_sqrt_iswap,
-)
-
-from cirq.transformers.transformer_primitives import (
-    map_moments,
-    map_operations,
-    map_operations_and_unroll,
-    merge_moments,
-    merge_operations,
-    unroll_circuit_op,
-    unroll_circuit_op_greedy_earliest,
-    unroll_circuit_op_greedy_frontier,
 )
